server/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
BingImageCreator==0.1.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
EdgeGPT==0.1.23
email-validator==1.3.1
fastapi==0.95.1
frozenlist==1.3.3
GoogleBard==0.0.7
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.0.4
openai==0.27.4
orjson==3.8.10
prompt-toolkit==3.0.38
pydantic==1.10.7
Pygments==2.15.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rich==13.3.4
sniffio==1.3.0
starlette==0.26.1
tqdm==4.65.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.1
yarl==1.8.2
